// packages/frameworks/mdx-docs/src/runner/types.ts
// Copyright (c) Qualcomm Technologies, Inc. and/or its subsidiaries.
// SPDX-License-Identifier: BSD-3-Clause-Clear

export type Scope = Record<string, any> & {
  /** scope used by import statement */
  import?: Record<string, any>
}

export type RunnerOptions = {
  /** the code to run */
  code: string
  /** globals that could be used in code */
  scope?: Scope
}